requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
python-amazon-sp-api==0.8.0
gspread
oauth2client
apscheduler
pandas
openpyxl
